Rework how context is stored and managed to make it available in controllers as well as views
