+++ lib/rbexy/component_context.rb
@@ -1,0 +1,18 @@
+module Rbexy
+  module ComponentContext
+    def rbexy_context
+      @rbexy_context ||= [{}]
+    end
+
+    def create_context(name, value)
+      rbexy_context.last[name] = value
+    end
+
+    def use_context(name)
+      index = rbexy_context.rindex { |c| c.has_key?(name) }
+      index ?
+        rbexy_context[index][name] :
+        raise(ContextNotFound, "no parent context `#{name}`")
+    end
+  end
+end
