+++ lib/rbexy/rails/engine.rb
@@ -9,18 +9,16 @@ module Rbexy
         ActionView::Template.register_template_handler(:rbx, template_handler)
 
         ActiveSupport.on_load :action_controller_base do
-          helper Rbexy::ViewContextHelper
-          helper_method :rbexy_component_provider
           include ControllerHelper
         end
 
         if defined?(ViewComponent)
-          ViewComponent::Base.include Rbexy::ViewContextHelper
+          ViewComponent::Base.include ViewContextHelper
         end
 
         Rbexy.configure do |config|
           require "rbexy/component_providers/rbexy_provider"
-          config.component_provider = Rbexy::ComponentProviders::RbexyProvider.new
+          config.component_provider = ComponentProviders::RbexyProvider.new
           config.template_paths << ::Rails.root.join("app", "components")
           config.enable_context = true
         end
