+++ spec/dummy/app/controllers/context_controller.rb
@@ -1,0 +1,5 @@
+class ContextController < ApplicationController
+  def index
+    create_context(:thing, "Hello context from controller")
+  end
+end
