+++ spec/dummy/app/views/context/index.rbx
@@ -1,0 +1,1 @@
+<Context.Child />
